MotherPL hit fix, Enemysimple death fix

+++ src/custom_datas.h
@@ -230,7 +230,7 @@ typedef enum{
 
 struct InvItem{
 	INVITEMTYPE itemtype;
-	UINT8 quantity;
+	UINT16 quantity;
 	UINT8 configured;
 };
